Géolocalisation de l'employé : - Détection des entrées et sorties dans la zone de la mission - notification à la base de données - Problème à résoudre : Dans la BD, quand un second employé pointe sur la même mission, le pointage du premier employé est écrasé.

+++ Virtual Time Clock - Client/Mission.swift
@@ -19,9 +19,9 @@ class Mission {
     let fin: Date
     let latitude: Double
     let longitude: Double
-    let rayon: Int
+    let rayon: Double
     
-    init(id: String, titre: String, lieu: String, description: String, debut: Date, fin: Date, latitude: Double, longitude: Double, rayon: Int) {
+    init(id: String, titre: String, lieu: String, description: String, debut: Date, fin: Date, latitude: Double, longitude: Double, rayon: Double) {
         self.id = id
         self.titre = titre
         self.lieu = lieu
